+++ CHANGELOG.rst
@@ -26,8 +26,12 @@ Changed
 - narwhalified ModeImputer `#321 <https://github.com/lvgig/tubular/issues/321>_`
 - fixed issues with all null and nullable-bool column handling in dataframe_init_dispatch
 - added NaN error handling to WeightColumnMixin
-- narwhalified MeanImputer `#344 https://github.com/lvgig/tubular/issues/344_`
-- narwhalified BaseNumericTransformer `#358 https://github.com/lvgig/tubular/issues/358`
+- narwhalified MeanImputer `#344 <https://github.com/lvgig/tubular/issues/344>_`
+- narwhalified BaseNumericTransformer `#358 <https://github.com/lvgig/tubular/issues/358>_`
+- narwhalified BaseCappingTransformer `#357 <https://github.com/lvgig/tubular/issues/357>_`
+- placeholder
+- placeholder
+- placeholder
 - placeholder
 - placeholder